+++ allshoppings-services/src/main/java/mobi/allshoppings/ranking/PointsService.java
@@ -18,9 +18,12 @@ public interface PointsService {
 	public static final int ACTION_COUPON_REDEEMED = 9;
 	public static final int ACTION_COUPON_REJECTED = 10;
 
-	void enqueue(final String userId, final Integer action, final Integer targetKind, final String targetId, final String data) throws ASException;
-	long calculatePointsForAction(final String userId, final Integer action, final Integer targetKind, final String targetId, String data);
-	long assignPointsForAction(final String userId, final Integer action, final Integer targetKind, final String targetId, String data) throws ASException;
+	void enqueue(final String userId, final Integer action, final byte targetKind,
+			final String targetId, final String data) throws ASException;
+	long calculatePointsForAction(final String userId, final Integer action,
+			final byte targetKind, final String targetId, String data);
+	long assignPointsForAction(final String userId, final Integer action, final byte targetKind,
+			final String targetId, String data) throws ASException;
 	long getTotalPointsForAction(final String userId, final Integer action) throws ASException;
 	long recountPointsForUser(final String userId, final Date fromDate, final Date toDate) throws ASException;
